requirements.txt
minio
requests
apache-airflow-providers-docker
apache-airflow-providers-slack
